backend/requirements.txt
fastapi
uvicorn
boto3
langchain
langchain-aws
langchain-core
langgraph
python-dotenv
pydantic
pydantic[email]
requests
PyJWT
email-validator
